chore: normalize mixins dir structure
