+++ mixins/withSelections/withSelections.mdx
@@ -1,8 +1,26 @@
+<!--
+ Copyright 2021 Comcast Cable Communications Management, LLC
+
+ Licensed under the Apache License, Version 2.0 (the "License");
+ you may not use this file except in compliance with the License.
+ You may obtain a copy of the License at
+
+ http://www.apache.org/licenses/LICENSE-2.0
+
+ Unless required by applicable law or agreed to in writing, software
+ distributed under the License is distributed on an "AS IS" BASIS,
+ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ See the License for the specific language governing permissions and
+ limitations under the License.
+
+ SPDX-License-Identifier: Apache-2.0
+ -->
+
 import { Canvas, Story } from '@storybook/addon-docs/blocks';
 
 # withSelections
 
-The `withSelections` mixin can be wrapped around a parent component that holds a list of components with a "selected" state where only one child component may have `isSelected=true` at a time. 
+The `withSelections` mixin can be wrapped around a parent component that holds a list of components with a "selected" state where only one child component may have `isSelected=true` at a time.
 
 <Canvas>
   <Story id="mixins-withselections--basic" />
@@ -150,7 +168,7 @@ Default value: `'items'`. Defines the name of the property applied to the wrappe
 ```js
 withSelections(FocusManager, {
   keys: {
-    items: 'children' 
+    items: 'children'
   }
 });
 ```
