+++ mixins/withTags/withTags.mdx
@@ -1,4 +1,5 @@
-<!-- Copyright 2020 Comcast Cable Communications Management, LLC
+<!--
+Copyright 2021 Comcast Cable Communications Management, LLC
 
 Licensed under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License.
@@ -12,8 +13,9 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 
-SPDX-License-Identifier: Apache-2.0 -->
- 
+SPDX-License-Identifier: Apache-2.0
+-->
+
 import { Canvas, Story } from '@storybook/addon-docs/blocks';
 
 # withTags
@@ -35,7 +37,7 @@ class Example extends withTags(lng.Component) {
   }
 
   static get tags() {
-    return ['Item', { name: 'Image', path: 'Item.Image'}];
+    return ['Item', { name: 'Image', path: 'Item.Image' }];
   }
 }
 ```
