--- mixins/withTheme.js
@@ -1,31 +1,0 @@
-/**
-* Copyright 2020 Comcast Cable Communications Management, LLC
-*
-* Licensed under the Apache License, Version 2.0 (the "License");
-* you may not use this file except in compliance with the License.
-* You may obtain a copy of the License at
-*
-* http://www.apache.org/licenses/LICENSE-2.0
-*
-* Unless required by applicable law or agreed to in writing, software
-* distributed under the License is distributed on an "AS IS" BASIS,
-* WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-* See the License for the specific language governing permissions and
-* limitations under the License.
-*
-* SPDX-License-Identifier: Apache-2.0
-*/
-import createTheme from '../components/Styles/createTheme';
-
-export default function withTheme(Base, theme) {
-  let defaultTheme = Base.theme;
-  const _theme = defaultTheme
-    ? createTheme(theme, defaultTheme)
-    : createTheme(theme);
-
-  return class extends Base {
-    static get name() { return Base.name }
-    static get theme() { return _theme };
-    get theme() { return _theme }
-  }
-}
