+++ mixins/withTransitions/withTransitions.mdx
@@ -1,5 +1,23 @@
+<!--
+ Copyright 2020 Comcast Cable Communications Management, LLC
+
+ Licensed under the Apache License, Version 2.0 (the "License");
+ you may not use this file except in compliance with the License.
+ You may obtain a copy of the License at
+
+ http://www.apache.org/licenses/LICENSE-2.0
+
+ Unless required by applicable law or agreed to in writing, software
+ distributed under the License is distributed on an "AS IS" BASIS,
+ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ See the License for the specific language governing permissions and
+ limitations under the License.
+
+ SPDX-License-Identifier: Apache-2.0
+ -->
+
 import { Canvas, Story } from '@storybook/addon-docs/blocks';
-import { DocsLink, ObjectFormat } from '../.storybook/utils';
+import { DocsLink, ObjectFormat } from '../../.storybook/utils';
 
 # withTransitions
 
@@ -8,7 +26,7 @@ Wrap a component with `withTransitions` to bulk apply transition settings to sta
 ```js
 import { withTransitions } from '@lightningjs/ui-components/mixins';
 
-class MyExample extends lng.Component {};
+class MyExample extends lng.Component {}
 
 export default withTransitions(MyExample);
 ```
@@ -21,18 +39,26 @@ export default withTransitions(MyExample);
 
 See the table below for available options:
 
-export const TransitionType = () => <ObjectFormat object={{ duration: 'number', delay: 'number', timingFunction: 'string' }} />
+export const TransitionType = () => (
+  <ObjectFormat
+    object={{ duration: 'number', delay: 'number', timingFunction: 'string' }}
+  />
+);
 
-export const TransitionDefault = () => <ObjectFormat object={{
-    duration: 0.3,
-    delay: 0,
-    timingFunction: 'cubic-bezier(0.80, 0.00, 0.20, 1.00)'
-  }} />
+export const TransitionDefault = () => (
+  <ObjectFormat
+    object={{
+      duration: 0.3,
+      delay: 0,
+      timingFunction: 'cubic-bezier(0.80, 0.00, 0.20, 1.00)'
+    }}
+  />
+);
 
-name|type|default|description
---|--|--|--
-properties|string[]|['alpha', 'color', 'scale', 'h', 'w', 'x', 'y']|names of properties to which transition settings should be applied
-transition|<TransitionType />|<TransitionDefault />|transition values to be applied. For more info, see <DocsLink id="lng.Transition">Transition docs</DocsLink>
+| name       | type               | default                                         | description                                                                                                  |
+| ---------- | ------------------ | ----------------------------------------------- | ------------------------------------------------------------------------------------------------------------ |
+| properties | string[]           | ['alpha', 'color', 'scale', 'h', 'w', 'x', 'y'] | names of properties to which transition settings should be applied                                           |
+| transition | <TransitionType /> | <TransitionDefault />                           | transition values to be applied. For more info, see <DocsLink id="lng.Transition">Transition docs</DocsLink> |
 
 ### How it works
 
@@ -49,7 +75,7 @@ class Container extends lng.Component {
       Child: {
         type: Child
       }
-    }
+    };
   }
 }
 ```
@@ -77,7 +103,7 @@ class Container extends lng.Component {
         }
       },
       Child: { type: Child }
-    }
+    };
   }
 
   _init() {
@@ -87,7 +113,7 @@ class Container extends lng.Component {
         timingFunction: 'ease-out',
         delay: 0 // applied by default since we ommitted it in the options
       }
-    }
+    };
   }
 }
 ```
@@ -101,15 +127,15 @@ class Container extends lng.Component {
   static _template() {
     return {
       Child1: { type: Child },
-      Child2: { type: Child },
-    }
+      Child2: { type: Child }
+    };
   }
 }
 
 export default withTransitions(Container, {
   properties: ['alpha'],
   transition: { duration: 0.5, timingFunction: 'ease-out' }
-})
+});
 ```
 
 If you want to override one of the global transitions:
@@ -125,14 +151,14 @@ class Container extends lng.Component {
           alpha: { delay: 0.3 }
         }
       }
-    }
+    };
   }
 }
 
 export default withTransitions(Container, {
   properties: ['alpha'],
   transition: { duration: 0.5, timingFunction: 'ease-out' }
-})
+});
 ```
 
 In the above example, `Container.transitions` and `Child1.transitions` will look like:
